chore(grpc): move metrics singletons to pkg/
