+++ central/telemetry/gatherers/gatherer_test.go
@@ -7,10 +7,10 @@ import (
 	"encoding/json"
 	"testing"
 
-	"github.com/stackrox/rox/central/grpc/metrics"
 	installationPostgres "github.com/stackrox/rox/central/installation/store/postgres"
 	"github.com/stackrox/rox/central/sensorupgradeconfig/datastore/mocks"
 	"github.com/stackrox/rox/generated/storage"
+	"github.com/stackrox/rox/pkg/grpc/metrics"
 	"github.com/stackrox/rox/pkg/postgres"
 	"github.com/stackrox/rox/pkg/postgres/pgtest"
 	"github.com/stackrox/rox/pkg/sac"
